webapp/webapp/requirements.txt
fastapi
uvicorn
python-multipart
jinja2
torch
torchvision
transformers
timm
pandas
numpy
albumentations
pillow
faiss-cpu